# src/docs/openapi-spec.yaml
openapi: 3.0.0
paths:
  /sale:
    patch:
      operationId: AmendSaleController_amendSale
      parameters: []
      requestBody:
        required: true
        description: Request body for amending a sales item
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  example: 2024-02-22T17:29:39Z
                invoiceId:
                  type: string
                  example: 3419027d-960f-4e8f-b8b7-f7b2b4791824
                itemId:
                  type: string
                  example: 02db47b6-fe68-4005-a827-24c6e962f3df
                cost:
                  type: number
                  example: 798
                taxRate:
                  type: number
                  example: 0.15
      responses:
        "202":
          description: The amendment has been successfully processed.
        "400":
          description: Invalid amendment request.
      summary: Amend a sales item
      tags:
        - sale
  /tax-position:
    get:
      operationId: TaxController_getTaxPosition
      parameters:
        - name: date
          required: true
          in: query
          description: Date and time in ISO 8601 format (e.g., 2024-02-22T17:29:39Z)
          schema:
            example: 2024-02-22T17:29:39Z
            type: string
      responses:
        "200":
          description: Tax position for the given date.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxDto"
        "400":
          description: Invalid or missing date parameter.
      summary: Get the tax position for a specific date
      tags:
        - tax-position
  /transactions:
    post:
      operationId: TransactionController_createTransaction
      parameters: []
      requestBody:
        required: true
        description: Request body for creating a transaction (SALES or TAX_PAYMENT)
        content:
          application/json:
            schema:
              oneOf:
                - example:
                    eventType: SALES
                    date: 2024-02-22T17:29:39Z
                    invoiceId: 3419027d-960f-4e8f-b8b7-f7b2b4791824
                    items:
                      - itemId: 02db47b6-fe68-4005-a827-24c6e962f3df
                        cost: 1099
                        taxRate: 0.2
                - example:
                    eventType: TAX_PAYMENT
                    date: 2024-02-22T17:29:39Z
                    amount: 74901
      responses:
        "202":
          description: The transaction has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionEntity"
        "400":
          description: Invalid transaction data.
      summary: Create a new transaction (SALES or TAX_PAYMENT)
      tags:
        - transactions
info:
  title: Novabook Tax Service
  description: API documentation for the Novabook Tax Service
  version: "1.0"
  contact: {}
tags: []
servers: []
components:
  schemas:
    TaxDto:
      type: object
      properties:
        date:
          type: string
          description: Date of the tax position in ISO 8601 format
          example: 2024-02-22T17:29:39Z
        taxPosition:
          type: number
          description: Calculated tax position at the given date in pennies
          example: 500
      required:
        - date
        - taxPosition
    TransactionEntity:
      type: object
      properties: {}